Refactor builtinExtensionTypes
Refactor builtinExtensionTypes to take enum ExtensionTypes
for better type safety and code clean-up.

The static cast to int is now an implementation detail.

All client-side static casts (to int or word) have been removed.

Based on Facebook D8959512

+++ ext/Objects/boolobject.cpp
@@ -69,7 +69,7 @@ extern "C" PyTypeObject* PyBool_Type_Ptr() {
   Thread* thread = Thread::currentThread();
   Runtime* runtime = thread->runtime();
   return static_cast<PyTypeObject*>(
-      runtime->builtinExtensionTypes(static_cast<int>(ExtensionTypes::kBool)));
+      runtime->builtinExtensionTypes(ExtensionTypes::kBool));
 }
 
 extern "C" PyObject* PyTrue_Ptr() {
